+++ Namco/solid_version/constants.py
